server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asttokens==2.4.1
attrs==24.3.0
Authlib==1.5.2
banks==2.1.2
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
filetype==1.2.0
flower==2.0.1
flupy==1.2.1
frozenlist==1.6.0
fsspec==2025.3.2
google-adk==0.3.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.90.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.12.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.0
graphviz==0.20.3
greenlet==3.2.1
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
langchain==0.3.24
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.38
linkify-it-py==2.0.3
litellm==1.67.2
llama-cloud==0.1.19
llama-cloud-services==0.6.16
llama-index==0.12.33
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.33.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.16
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.6.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numpy==2.2.5
openai==1.74.0
openpipe==4.50.0
opentelemetry-api==1.32.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
postgrest==1.0.1
progressbar2==4.5.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic-yaml==1.5.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.4.0
pyright==1.1.400
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.11
ruamel.yaml.clib==0.2.12
ruff==0.11.9
selenium==4.31.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
striprtf==0.0.26
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
textual==3.2.0
tiktoken==0.9.0
tokenizers==0.21.1
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vecs==0.4.5
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
